### 2022-11-27 #4974 @editorialbot
[PRE REVIEW]: PyExperimenter: Easily distribute experiments and track results
**Submitting author:** @tornede (<a href="http://orcid.org/0000-0001-9954-462X">Tanja Tornede</a>)
**Repository:** https://github.com/tornede/py_experimenter
**Branch with paper.md** (empty if default branch): develop
**Version:** 1.1.0
