+++ paper/paper.bib
@@ -180,7 +180,8 @@ author = {ClearML},
   booktitle = {Proceedings of the 16th python in science conference},
   volume    = {28},
   pages     = {49--56},
-  year      = {2017}
+  year      = {2017}, 
+  doi       = {10.25080/shinma-7f4c6e7-008}
 }
 
 @misc{guildai,
