### 2022-12-14 #4974 @arfon
@tornede – apologies for the slow progress here – thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
