### 2023-01-02 c208351
Reorder functionality in utils.py
Now the functions are ordered by importance.
