### 2023-01-02 d757a2e
Creation of Logtables.
Logtables now get created in 'create_table_if_not_existing'.

Additionally the code for creating tables was heavily reworked.
