### 2023-01-03 672b624
Add execute_queries and testcases
