### 2023-01-03 fc1ca22
Add shorthand notation.
