+++ test/test_logtables/test_mysql.py
@@ -62,7 +62,7 @@ def test_logtable_insertion(database_connector_mock):
 def test_delete_logtable(execution_mock, close_connection_mock, commit_mocck, fetchall_mock, cursor_mock, connect_mock, test_connection_mock, create_database_mock):
     fetchall_mock.return_value = cursor_mock.return_value = connect_mock.return_value = commit_mocck.return_value = None
     close_connection_mock.return_value = test_connection_mock.return_value = create_database_mock.return_value = execution_mock.return_value = None
-    experimenter = PyExperimenter(os.path.join('test', 'test_logtables', 'mysql_logtables.cfg'))  # todo fill this
+    experimenter = PyExperimenter(os.path.join('test', 'test_logtables', 'mysql_logtables.cfg')) 
     experimenter.delete_table()
     execution_mock.assert_has_calls([call(None, 'DROP TABLE IF EXISTS test_mysql_log'),
                                      call(None, 'DROP TABLE IF EXISTS test_mysql_log2'),
