### 2023-01-09 347b8bd
Modify logtable naming.
