+++ test/test_logtables/test_sqlite.py
@@ -32,7 +32,7 @@ def test_tables_created(execute_mock, close_connection_mock, fetchall_mock, curs
                                                 'start_date VARCHAR(255) DEFAULT NULL,name LONGTEXT DEFAULT NULL,machine VARCHAR(255) DEFAULT NULL,'
                                                 'sin VARCHAR(255) DEFAULT NULL,cos VARCHAR(255) DEFAULT NULL,end_date VARCHAR(255) DEFAULT NULL,'
                                                 'error LONGTEXT DEFAULT NULL);')
-    assert execute_mock.mock_calls[1][1][1] == ('CREATE TABLE test_sqlite_log (ID INTEGER PRIMARY KEY AUTOINCREMENT, test int DEFAULT NULL,'
+    assert execute_mock.mock_calls[1][1][1] == ('CREATE TABLE test_sqlite_logtables__test_sqlite_log (ID INTEGER PRIMARY KEY AUTOINCREMENT, test int DEFAULT NULL,'
                                                 'experiment_id INTEGER, FOREIGN KEY (experiment_id) REFERENCES test_sqlite_logtables(ID) ON DELETE CASCADE);')
 
 
@@ -41,12 +41,13 @@ def test_logtable_insertion(database_connector_mock):
     config = ConfigParser()
     config.read(os.path.join('test', 'test_logtables', 'sqlite_logtables.cfg'))
     result_processor = ResultProcessor(config, None, None, None, 0)
+    result_processor._table_name = 'table_name'
     result_processor.process_logs({'test_table_0': {'test0': 'test', 'test1': 'test'},
                                    'test_table_1': {'test0': 'test'}})
     result_processor._dbconnector.execute_queries.assert_called()
     result_processor._dbconnector.execute_queries.assert_called_with(
-        ['INSERT INTO test_table_0 (test0, test1, experiment_id) VALUES (test, test, 0)',
-         'INSERT INTO test_table_1 (test0, experiment_id) VALUES (test, 0)'])
+        ['INSERT INTO table_name__test_table_0 (test0, test1, experiment_id) VALUES (test, test, 0)',
+         'INSERT INTO table_name__test_table_1 (test0, experiment_id) VALUES (test, 0)'])
 
 
 @patch('py_experimenter.experimenter.DatabaseConnectorLITE._test_connection')
@@ -61,7 +62,8 @@ def test_delete_logtable(execution_mock, close_connection_mock, commit_mocck, fe
     close_connection_mock.return_value = test_connection_mock.return_value = execution_mock.return_value = None
     experimenter = PyExperimenter(os.path.join('test', 'test_logtables', 'sqlite_logtables.cfg'))
     experimenter.delete_table()
-    execution_mock.assert_has_calls([call(None, 'DROP TABLE IF EXISTS test_sqlite_log'),call(None, 'DROP TABLE IF EXISTS test_sqlite_log2'), call(None, 'DROP TABLE IF EXISTS test_sqlite_logtables')])
+    execution_mock.assert_has_calls([call(None, 'DROP TABLE IF EXISTS test_sqlite_logtables__test_sqlite_log'),
+                                     call(None, 'DROP TABLE IF EXISTS test_sqlite_logtables__test_sqlite_log2'), call(None, 'DROP TABLE IF EXISTS test_sqlite_logtables')])
 
 
 ####################### Integration Test
@@ -88,10 +90,10 @@ def test_integration():
     experimenter.execute(own_function, 1)
     connection = experimenter.dbconnector.connect()
     cursor = experimenter.dbconnector.cursor(connection)
[…]

[…]
