### 2023-01-12 fba7a46
Changed supported python versions (#86)
* Updated pyproject.toml to support Python >= 3.9
* Updated tests to be performed for Python >= 3.9
