+++ pyproject.toml
@@ -28,7 +28,7 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.7"
+python = "^3.9"
 numpy = ">=1.15"
 pandas = ">=1.0"
 mysql-connector-python = ">=8.0"
