+++ _sources/autoapi/py_experimenter/database_connector_mysql/index.rst.txt
@@ -21,6 +21,9 @@ Classes
 
    Bases: :py:obj:`py_experimenter.database_connector.DatabaseConnector`
 
+   Helper class that provides a standard way to create an ABC using
+   inheritance.
+
    .. py:attribute:: _write_to_database_separator
       :annotation: = ', '
 
