### 2023-01-12 e99de27
Extracted string for ID condition
