+++ _sources/autoapi/py_experimenter/result_processor/index.rst.txt
@@ -42,7 +42,7 @@ Attributes
 
    
 
-.. py:class:: ResultProcessor(_config: dict, credential_path, table_name: str, condition: dict, result_fields: List[str])
+.. py:class:: ResultProcessor(_config: dict, credential_path, table_name: str, result_fields: List[str], experiment_id: int)
 
    Class for processing the results from an experiment. Use this class whenever you want to write results to the
    database.
@@ -68,9 +68,6 @@ Attributes
 
 
    .. py:method:: _set_name(name)
-
-
-   .. py:method:: _not_executed_yet() -> bool
 
 
    .. py:method:: _valid_result_fields(result_fields)
