### 2023-01-16 9acddc4
Fix order of columns in the logtables.
