### 2023-01-16 1b9799f
Add timestamps to the result_tables
