+++ test/test_logtables/test_mysql.py
@@ -2,6 +2,7 @@ import os
 from configparser import ConfigParser
 from math import cos, sin
 
+from freezegun import freeze_time
 from mock import call, patch
 
 from py_experimenter.database_connector import DatabaseConnector
@@ -34,12 +35,15 @@ def test_tables_created(execute_mock, close_connection_mock, fetchall_mock, curs
                                                 'start_date DATETIME DEFAULT NULL,name LONGTEXT DEFAULT NULL,machine VARCHAR(255) DEFAULT NULL,'
                                                 'sin VARCHAR(255) DEFAULT NULL,cos VARCHAR(255) DEFAULT NULL,end_date DATETIME DEFAULT NULL,'
                                                 'error LONGTEXT DEFAULT NULL);')
-    assert execute_mock.mock_calls[2][1][1] == ('CREATE TABLE test_mysql_logtables__test_mysql_log (ID INTEGER PRIMARY KEY AUTO_INCREMENT, experiment_id INTEGER,'
-                                                ' test int DEFAULT NULL, FOREIGN KEY (experiment_id) REFERENCES test_mysql_logtables(ID) ON DELETE CASCADE);')
+    assert execute_mock.mock_calls[2][1][1] == ('CREATE TABLE test_mysql_logtables__test_mysql_log (ID INTEGER PRIMARY KEY AUTO_INCREMENT,'
+                                                ' experiment_id INTEGER, timestamp DATETIME, test int DEFAULT NULL, FOREIGN KEY (experiment_id)'
+                                                ' REFERENCES test_mysql_logtables(ID) ON DELETE CASCADE);')
 
 
+@freeze_time("2012-01-14 03:21:34")
 @patch('py_experimenter.result_processor.DatabaseConnectorMYSQL')
 def test_logtable_insertion(database_connector_mock):
+    fixed_time = '2012-01-14 03:21:34'
     config = ConfigParser()
     config.read(os.path.join('test', 'test_logtables', 'mysql_logtables.cfg'))
     result_processor = ResultProcessor(config, None, None, None, 0)
@@ -48,8 +52,8 @@ def test_logtable_insertion(database_connector_mock):
                                    'test_table_1': {'test0': 'test'}})
     result_processor._dbconnector.execute_queries.assert_called()
     result_processor._dbconnector.execute_queries.assert_called_with(
-        ['INSERT INTO some_table_name__test_table_0 (test0, test1, experiment_id) VALUES (test, test, 0)',
-         'INSERT INTO some_table_name__test_table_1 (test0, experiment_id) VALUES (test, 0)'])
+        [f'INSERT INTO some_table_name__test_table_0 (test0, test1, experiment_id, timestamp) VALUES (test, test, 0, \'{fixed_time}\')',
+         f'INSERT INTO some_table_name__test_table_1 (test0, experiment_id, timestamp) VALUES (test, 0, \'{fixed_time}\')'])
 
 
 @patch('py_experimenter.experimenter.DatabaseConnectorMYSQL._create_database_if_not_existing')
@@ -63,14 +67,14 @@ def test_logtable_insertion(database_connector_mock):
 def test_delete_logtable(execution_mock, close_connection_mock, commit_mocck, fetchall_mock, cursor_mock, connect_mock, test_connection_mock, create_database_mock):
     fetchall_mock.return_value = cursor_mock.return_value = connect_mock.return_value = commit_mocck.return_value = None
     close_connection_mock.return_value = test_connection_mock.return_value = create_database_mock.return_value = execution_mock.return_value = None
[…]
