### 2023-01-16 73b8167
Unify timestamp handling.
Move timestamp handling to database_connector whenever possible.
Introduce utils.get_timestamp_representation to remove duplicated code.
