+++ py_experimenter/result_processor.py
@@ -1,9 +1,9 @@
 import logging
 from copy import deepcopy
-from datetime import datetime
 from typing import Dict, List, Tuple
 
 import py_experimenter.utils as utils
+from py_experimenter.database_connector import DatabaseConnector
 from py_experimenter.database_connector_lite import DatabaseConnectorLITE
 from py_experimenter.database_connector_mysql import DatabaseConnectorMYSQL
 from py_experimenter.exceptions import InvalidConfigError, InvalidResultFieldError
@@ -31,11 +31,11 @@ class ResultProcessor:
         self._timestamp_on_result_fields = utils.timestamps_for_result_fields(self._config)
         self._experiment_id = experiment_id
         self._experiment_id_condition = f'ID = {experiment_id}'
-        
+
         if _config['PY_EXPERIMENTER']['provider'] == 'sqlite':
-            self._dbconnector = DatabaseConnectorLITE(_config)
+            self._dbconnector: DatabaseConnector = DatabaseConnectorLITE(_config)
         elif _config['PY_EXPERIMENTER']['provider'] == 'mysql':
-            self._dbconnector = DatabaseConnectorMYSQL(_config, credential_path)
+            self._dbconnector: DatabaseConnector = DatabaseConnectorMYSQL(_config, credential_path)
         else:
             raise InvalidConfigError("Invalid database provider!")
 
@@ -45,20 +45,20 @@ class ResultProcessor:
         want to write results to the database.
         :param results: Dictionary with result field name and result value pairs.
         """
-        time = datetime.now().strftime("%m/%d/%Y, %H:%M:%S")
         if not self._valid_result_fields(list(results.keys())):
             invalid_result_keys = set(list(results.keys())) - set(self._result_fields)
             raise InvalidResultFieldError(f"Invalid result keys: {invalid_result_keys}")
 
         if self._timestamp_on_result_fields:
-            results = self.__class__._add_timestamps_to_results(results, time)
+            results = self.__class__._add_timestamps_to_results(results)
 
         keys = self._dbconnector.escape_sql_chars(*list(results.keys()))
         values = self._dbconnector.escape_sql_chars(*list(results.values()))
         self._dbconnector._update_database(keys=keys, values=values, where=self._experiment_id_condition)
 
     @staticmethod
-    def _add_timestamps_to_results(results: dict, time: datetime) -> List[Tuple[str, object]]:
+    def _add_timestamps_to_results(results: dict) -> List[Tuple[str, object]]:
+        time = utils.get_timestamp_representation()
         result_fields_with_timestep = deepcopy(results)
         for result_field, value in sorted(results.items()):
             result_fields_with_timestep[result_field] = value
@@ -67,7 +67,7 @@ class ResultProcessor:
 
     def process_logs(self, logs: Dict[str, Dict[str, str]]) -> None:
         queries = []
-        time  = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+        time = utils.get_timestamp_representation()
         for logtable_identifier, log_entries in logs.items():
             logtable_name = f'{self._table_name}__{logtable_identifier}'
             log_entries['experiment_id'] = str(self._experiment_id)
@@ -76,9 +76,8 @@ class ResultProcessor:
                 f"INSERT INTO {logtable_name} ({', '.join(log_entries.keys())}) VALUES ({', '.join(map(lambda x: str(x), log_entries.values()))})")
         self._dbconnector.execute_queries(queries)
 
-    def _change_status(self, status):
-        time = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-        
+    def _change_status(self, status:str):
+        time = utils.get_timestamp_representation()
         if status == 'done' or status == 'error':
[…]
