### 2023-01-16 c459e5b
Add freezegun to poetry environment.
The dependency is used for several tests.
