### 2023-01-17 cbfc128
Change used mysql-connector implementation to c.
