### 2023-01-20 fcfb2a0
Use prepared statement to escape values
- delete old escaping functionality
- add utils.get_current_time to avoid duplicated code
- database_connector.update_database now takes different parameters
- reformatting of test_run_sqlite_experiments
- add testcase to check if errors get written into the database correct
