+++ py_experimenter/result_processor.py
@@ -45,7 +45,7 @@ class ResultProcessor:
         want to write results to the database.
         :param results: Dictionary with result field name and result value pairs.
         """
-        time = datetime.now().strftime("%m/%d/%Y, %H:%M:%S")
+        time = utils.get_current_time()
         if not self._valid_result_fields(list(results.keys())):
             invalid_result_keys = set(list(results.keys())) - set(self._result_fields)
             raise InvalidResultFieldError(f"Invalid result keys: {invalid_result_keys}")
@@ -53,9 +53,7 @@ class ResultProcessor:
         if self._timestamp_on_result_fields:
             results = self.__class__._add_timestamps_to_results(results, time)
 
-        keys = self._dbconnector.escape_sql_chars(*list(results.keys()))
-        values = self._dbconnector.escape_sql_chars(*list(results.values()))
-        self._dbconnector._update_database(keys=keys, values=values, where=self._experiment_id_condition)
+        self._dbconnector.update_database(self._table_name, values=results, condition=self._experiment_id_condition)
 
     @staticmethod
     def _add_timestamps_to_results(results: dict, time: datetime) -> List[Tuple[str, object]]:
@@ -66,23 +64,19 @@ class ResultProcessor:
         return result_fields_with_timestep
 
     def _change_status(self, status):
-        time = datetime.now()
-        time = time.strftime("%m/%d/%Y, %H:%M:%S")
+        values = {'status': status,
+                  'end_date': utils.get_current_time()}
 
-        if status == 'done' or status == 'error':
-            self._dbconnector._update_database(keys=['status', 'end_date'], values=[status, time], where=self._experiment_id_condition)
+        self._dbconnector._update_database(keys=['status', 'end_date'], values=[status, time], where=f'ID = {self._experiment_id}')
 
     def _write_error(self, error_msg):
-        self._dbconnector._update_database(keys=['error'], values=[error_msg], where=self._experiment_id_condition)
+        self._dbconnector._update_database(keys=['error'], values=[error_msg], where=f'ID = {self._experiment_id}')
 
     def _set_machine(self, machine_id):
-        self._dbconnector._update_database(keys=['machine'], values=[machine_id], where=self._experiment_id_condition)
+        self._dbconnector._update_database(keys=['machine'], values=[machine_id], where=f'ID = {self._experiment_id}')
 
     def _set_name(self, name):
-        self._dbconnector._update_database(keys=['name'], values=[name], where=self._experiment_id_condition)
-
-    def _not_executed_yet(self) -> bool:
-        return self._dbconnector.not_executed_yet(where=self._experiment_id_condition)
+        self._dbconnector._update_database(keys=['name'], values=[name], where=f'ID = {self._experiment_id}')
 
     def _valid_result_fields(self, result_fields):
         return set(result_fields).issubset(set(self._result_fields))
