+++ py_experimenter/database_connector.py
@@ -1,6 +1,5 @@
 import abc
 import logging
-from datetime import datetime
 from typing import Dict, List, Optional, Tuple, Union
 
 import pandas as pd
