### 2023-01-20 c28c1d2
Adapt testcase to be run on VM
