+++ test/test_run_experiments/test_run_sqlite_experiment.py
@@ -109,9 +109,10 @@ def test_run_error_experiment():
     assert entries[0][7] == socket.gethostname()
     assert entries[0][8] == None
     assert entries[0][9] == None
+    first_cwd_mention = os.getcwd() + os.sep + os.path.join('py_experimenter', 'experimenter.py')
+    second_cwd_mention = os.getcwd() + os.sep + os.path.join('test', 'test_run_experiments', 'test_run_sqlite_experiment.py')
     assert entries[0][11] == ('Traceback (most recent call last):\n  File'
-                              f' "{os.getcwd()}/py_experimenter/experimenter.py"'
+                              f' "{first_cwd_mention}"'
                               ', line 403, in _execution_wrapper\n    experiment_function(keyfield_values, result_processor,'
-                              f' custom_fields)\n  File "{os.getcwd()}/test/'
-                              'test_run_experiments/test_run_sqlite_experiment.py", line 82, in error_function\n    '
+                              f' custom_fields)\n  File "{second_cwd_mention}", line 82, in error_function\n    '
                               'raise Exception("Error with weird symbos \'@#$%&/\\()=")\nException: Error with weird symbos \'@#$%&/\\()=\n')
