### 2023-02-10 #4974 @timtroendle
Hi @ArsamAryandoust, sorry for being unclear on this. You will receive clearer instructions once the review started. There will be a checklist for you to go through and you can then send smaller comments in the (yet-to-be-opened) review issue or open issues in the repository of PyExperimenter.

All of this will be done once the review started. Before, I am looking for at least one other reviewer.
