### 2023-02-13 #5149 @editorialbot
[REVIEW]: PyExperimenter: Easily distribute experiments and track results
**Submitting author:** @tornede (<a href="http://orcid.org/0000-0001-9954-462X">Tanja Tornede</a>)
**Repository:** https://github.com/tornede/py_experimenter
**Branch with paper.md** (empty if default branch): develop
**Version:** v1.2.0
