### 2023-02-23 56fb2ce
Add get_logtable functionality.
