+++ docs/source/examples/index.rst
@@ -9,5 +9,5 @@ Examples
 
    ./example_general_usage.ipynb
    ./example_conditional_grid.ipynb
-
+   ./example_logtables
 
