+++ test/test_run_experiments/test_run_sqlite_experiment.py
@@ -3,6 +3,7 @@ import os
 import socket
 from math import cos, sin
 
+import pandas as pd
 from mysql.connector.errors import ProgrammingError
 
 from py_experimenter.experimenter import PyExperimenter
@@ -114,5 +115,45 @@ def test_run_error_experiment():
     assert entries[0][11] == ('Traceback (most recent call last):\n  File'
                               f' "{first_cwd_mention}"'
                               ', line 403, in _execution_wrapper\n    experiment_function(keyfield_values, result_processor,'
-                              f' custom_fields)\n  File "{second_cwd_mention}", line 82, in error_function\n    '
+                              f' custom_fields)\n  File "{second_cwd_mention}", line 83, in error_function\n    '
                               'raise Exception("Error with weird symbos \'@#$%&/\\()=")\nException: Error with weird symbos \'@#$%&/\\()=\n')
+
+
+def own_function_raising_errors(keyfields: dict, result_processor: ResultProcessor, custom_fields: dict):
+    error_code = keyfields['error_code']
+
+    # give list of special characters that frequently cause problems
+    characters = ['"', "'", '@', '#', '$', '%', '&', '/', '\\', '(', ')', '=', "`", "`some_text`", "^"]
+    if error_code == 0:
+        raise Exception("Error with weird symbos" + "".join(characters))
+    elif error_code == 1:
+        raise LookupError("Error with weird symbos" + str(characters))
+    elif error_code == 2:
+        raise ProgrammingError("Error with weird symbos" + str(characters))
+
+
+def test_raising_error_experiment():
+    experimenter = PyExperimenter(experiment_configuration_file_path=os.path.join('test', 'test_run_experiments', 'test_run_mysql_error_config.cfg'),
+                                  name='name')
+
+    try:
+        experimenter.delete_table()
+    except ProgrammingError as e:
+        logging.warning(e)
+
+    experimenter.fill_table_from_config()
+    experimenter.execute(own_function_raising_errors, -1)
+    table = experimenter.get_table()
+    'lala'
+    table = table[['ID', 'error_code', 'status', 'name']]
+    pd.testing.assert_frame_equal(
+        table,
+        pd.DataFrame(
+            {
+                'ID': [1, 2, 3],
+                'error_code': [0., 1., 2.],
+                'status': ['error', 'error', 'error'],
+                'name': ['name', 'name', 'name'],
+            }
+        )
+    )
