+++ test/test_run_experiments/test_run_mysql_experiment.py
@@ -2,6 +2,7 @@ import logging
 import os
 from math import cos, sin
 
+import pandas as pd
 from mysql.connector.errors import ProgrammingError
 
 from py_experimenter.experimenter import PyExperimenter
@@ -108,8 +109,51 @@ def test_run_error_experiment():
     assert entries[0][8] == None
     assert entries[0][9] == None
     assert entries[0][11] == ('Traceback (most recent call last):\n  File'
-                              ' "/home/tornede/remote_development/py_experimenter/py_experimenter/experimenter.py"'
+                              f' "{os.path.join(os.getcwd(),"py_experimenter","experimenter")}.py"'
                               ', line 403, in _execution_wrapper\n    experiment_function(keyfield_values, result_processor,'
-                              ' custom_fields)\n  File "/home/tornede/remote_development/py_experimenter/test/'
-                              'test_run_experiments/test_run_mysql_experiment.py", line 80, in error_function\n    '
+                              f' custom_fields)\n  File "{os.path.join(os.getcwd(), "test", "test_run_experiments", "test_run_mysql_experiment")}'
+                              '.py", line 81, in error_function\n    '
                               'raise Exception("Error with weird symbos \'@#$%&/\\()=")\nException: Error with weird symbos \'@#$%&/\\()=\n')
+
+
+def own_function_raising_errors(keyfields: dict, result_processor: ResultProcessor, custom_fields: dict):
+    error_code = keyfields['error_code']
+
+    # give list of special characters that frequently cause problems
+    characters = ['"', "'", '@', '#', '$', '%', '&', '/', '\\', '(', ')', '=', "`", "`some_text`", "^"]
+    if error_code == 0:
+        raise Exception("Error with weird symbos" + "".join(characters))
+    elif error_code == 1:
+        raise LookupError("Error with weird symbos" + "".join(characters))
+    elif error_code == 2:
+        raise ProgrammingError("Error with weird symbos" + "".join(characters))
+
+
+def test_raising_error_experiment():
+    experimenter = PyExperimenter(experiment_configuration_file_path=os.path.join('test', 'test_run_experiments', 'test_run_mysql_error_config.cfg'),
+                                  name='name')
+
+    try:
+        experimenter.delete_table()
+    except ProgrammingError as e:
+        logging.warning(e)
+
+    experimenter.fill_table_from_config()
+    experimenter.execute(own_function_raising_errors, -1)
+    table = experimenter.get_table()
+    'lala'
+    table = table[['ID', 'error_code', 'status', 'name']]
+    pd.testing.assert_frame_equal(
+        table,
+        pd.DataFrame(
+            {
+                'ID': [1, 2, 3],
+                'error_code': [0., 1., 2.],
+                'status': ['error', 'error', 'error'],
+                'name': ['name', 'name', 'name'],
+            }
+        )
+    )
+
+    
+    
