### 2023-03-12 tornede/py_experimenter#90 @schnorr
`reset_experiments` documention
Hi, the [`reset_experiments` documentation](https://tornede.github.io/py_experimenter/autoapi/py_experimenter/experimenter/index.html#py_experimenter.experimenter.PyExperimenter.reset_experiments) mention `states` as parameter, which is in disagreement with the [general usage documentation](https://tornede.github.io/py_experimenter/examples/example_general_usage.html#Restart-Failed-Experiments) that uses the `status` parameter. After installing pyexperimenter using the `pip` command, and attempting to replay the commands, it seems that neither `status` or `states` are accepted as parameter, I keep getting a `got an unexpected keyword argument` exception. The [current source code](https://github.com/tornede/py_experimenter/blob/main/py_experimenter/experimenter.py#L412) indicates `states`, in disagreement with the documentation. It only works when providing the argument without the parameter name.
