### 2023-03-12 #5149 @schnorr
So, just to let you know that I've completed my review. My opinion after regarding the source code, documentation, testing the commands suggested by the authors, etc, is that the `PyExperimenter` software has good maturity. It also has some people already using it, as can be seen in the Issues tabs for people requesting new functionality. I've only found a minor problem that I've registered as [an issue in their repository](https://github.com/tornede/py_experimenter/issues/90#issue-1620357634).

About the Functionality, I think the tool is very useful because usually R or Python packages are focused on generating the experimental project without handling the experiment run. The PyExperimenter let the user define a function that will receive the configuration parameters. The function must then run the experiment itself with those parameters and let PyExperimenter know that everything went okay. So, the functionality provides a way to define the experiment project and also run it (eventually in parallel with the `n_jobs` parameter). Of course, some R packages especially are much more advanced because they would allow one to define specific screening projects to optimize the number of experiments to be run (avoiding a cartesian product of all configurations - which might be overwhelming). But that is not the focus of this project, which is focused in a specific type of experimental project.

I've checked all the bullet points in the review, running `sloccount` (for the number of lines ~ about 1k lines), checking the number of committers, and so on. As consequence, I conclude that the software checks all the requirements. So, my general opinion is a go for accept.

Thank you once again for considering my for JOSS @timtroendle !
