### 2023-03-14 #5149 @ArsamAryandoust
The authors provide solid work that includes both extensive unit and integration
tests. A list of dependencies is available in the pyproject.toml file, maybe
could also be stated in the readme or a requirements.txt file (?). Overall, I
suggest to go for an accept only after addressing the following issues.

1. The documentations under docs/source/examples/ need to be improved. Test your
notebooks more accurately and make sure they run. I made exemplar pull requests.
Everything else works flawlessly.

2. Seeking support and creating issues is described under documents tab 'Need help?'.
Descriptions on how to contribute can be added.

3. I did not test the full functionality of the presented tool. However, the provided
toy examples are functional and seem to demonstrate the core functionality.

4. The paper can have small improvements. @Editor: please provide a proper PDF
file or other manuscript file for suggesting changes.

I like working with Docker. So I create a minimalistic Dockerfile:
```
cd
mkdir py_experimenter
cd py_experimenter
```
```
FROM ubuntu:latest
RUN apt-get update && apt-get upgrade -y
RUN apt-get install python3-pip -y
RUN pip3 install py-experimenter
RUN pip3 install scikit-learn

WORKDIR /py_experimenter

```

I build and enter the container:

```
docker build Docker -t py_experimenter_docker
docker run -it -v ~/py_experimenter_test:/py_experimenter py_experimenter_docker
```

Everything works flawlessly. In example notebooks, there are small issues. In code
section 1, insert the following command:
```
if not os.path.isdir('config'):
  os.mkdir('config')
experiment_configuration_file_path = os.path.join('config', 'example_general_usage.cfg')
with open(experiment_configuration_file_path, "w") as f:
  f.write(content)
```

In code snippet under 'Fill Table' section, change code into the following, to
ensure that code runs while also illustrating the example. I guess everyone will
still get the point also with dataset being iris again:
```
experimenter.fill_table_from_config()

experimenter.fill_table_with_rows(rows=[
      {'dataset': 'iris', 'cross_validation_splits': 3, 'seed': 42, 'kernel':'linear'}])

# showing database table
experimenter.get_table()
```

In code section under 'Restart Failed Experiments', change code into:
```
experimenter.reset_experiments('error')

# showing database table
experimenter.get_table()
```
