### 2023-03-14 #5149 @timtroendle
Thanks a lot for your extensive reviews, @schnorr and @ArsamAryandoust .

@ArsamAryandoust , the paper is available as [Markdown source](https://github.com/tornede/py_experimenter/blob/main/paper/paper.md). Would that help?
