### 2023-03-15 #5149 @ArsamAryandoust
@timtroendle thanks for the reference to the paper markdown. I am done with suggested changes to the paper and created these as a pull request to the markdown file.

The paper is well written and together with the documentation of the tool, it contains everything a user needs to know. I further believe the documentation is not perfect but a great initial version for a tool that irons out with more user feedback.
