+++ py_experimenter/database_connector.py
@@ -1,4 +1,5 @@
 import abc
+from datetime import datetime
 import logging
 from typing import Dict, List, Optional, Tuple, Union
 
@@ -161,9 +162,9 @@ class DatabaseConnector(abc.ABC):
     def _get_existing_rows(self, column_names):
         pass
 
-    def get_experiment_configuration(self, random_order: bool):
+    def get_experiment_configuration(self):
         try:
-            experiment_id, description, values = self._pull_open_experiment(random_order)
+            experiment_id, description, values = self._pull_open_experiment()
         except IndexError as e:
             raise NoExperimentsLeftException("No experiments left to execute")
         except Exception as e:
@@ -171,12 +172,10 @@ class DatabaseConnector(abc.ABC):
 
         return experiment_id, dict(zip([i[0] for i in description], *values))
 
-    def _execute_queries(self, connection, cursor, random_order) -> Tuple[int, List, List]:
-        if random_order:
-            order_by = self.__class__.random_order_string()
-        else:
-            order_by = "id"
-        time = utils.get_current_time()
+    def _execute_queries(self, connection, cursor) -> Tuple[int, List, List]:
+        order_by = "id"
+        time = datetime.now()
+        time = time.strftime("%m/%d/%Y, %H:%M:%S")
 
         self.execute(cursor, f"SELECT id FROM {self.table_name} WHERE status = 'created' ORDER BY {order_by} LIMIT 1;")
         experiment_id = self.fetchall(cursor)[0][0]
@@ -189,12 +188,8 @@ class DatabaseConnector(abc.ABC):
         description = cursor.description
         return experiment_id, description, values
 
-    @abc.abstractstaticmethod
-    def random_order_string():
-        pass
-
     @abc.abstractmethod
-    def _pull_open_experiment(self, random_order) -> Tuple[int, List, List]:
+    def _pull_open_experiment(self) -> Tuple[int, List, List]:
         pass
 
     def _write_to_database(self, keys, values) -> None:
