+++ py_experimenter/database_connector_lite.py
@@ -30,11 +30,11 @@ class DatabaseConnectorLITE(DatabaseConnector):
         except Error as err:
             raise DatabaseConnectionError(err)
 
-    def _pull_open_experiment(self, random_order) -> Tuple[int, List, List]:
+    def _pull_open_experiment(self) -> Tuple[int, List, List]:
         with connect(**self.database_credentials) as connection:
             try:
                 cursor = self.cursor(connection)
-                experiment_id, description, values = self._execute_queries(connection, cursor, random_order)
+                experiment_id, description, values = self._execute_queries(connection, cursor)
             except Exception as err:
                 connection.rollback()
                 raise err
@@ -56,10 +56,6 @@ class DatabaseConnectorLITE(DatabaseConnector):
         columns = self._exclude_fixed_columns([k[1] for k in self.fetchall(cursor)])
         config_columns = [k[0] for k in typed_fields]
         return set(columns) == set(config_columns)
-
-    @staticmethod
-    def random_order_string():
-        return 'RANDOM()'
 
     def _get_existing_rows(self, column_names):
         def _remove_string_markers(row):
