### 2023-03-28 da14fb8
Update docstring according to feedback
