### 2023-03-29 75260cc
Removed line testing the machine name
