### 2023-03-29 8809967
Bugfix logging of stack traces into mysql (#87)
* Change row identification in ResultProcessor. Instead of checking keyfields the experiment_id is utilized
* Change used mysql-connector implementation to c
* Use prepared statement to escape values
  - delete old escaping functionality
  - add utils.get_current_time to avoid duplicated code
  - database_connector.update_database now takes different parameters
  - reformatting of test_run_sqlite_experiments
  - add testcase to check if errors get written into the database correct
* Make tests Windows compatible
