+++ .github/workflows/tests.yml
@@ -26,7 +26,7 @@ jobs:
         with:
           virtualenvs-create: true
           virtualenvs-in-project: true
-      - name: disable modern installation
+      - name: Disable Modern Dependency Installation
         id: poetry-disable-modern
         run: poetry config installer.modern-installation false
       - name: Load cached venv
