### 2023-03-31 e7369d9
Add python script examples exacutable for windows
