+++ test/test_run_experiments/test_run_sqlite_experiment.py
@@ -114,7 +114,7 @@ def test_run_error_experiment():
     second_cwd_mention = os.getcwd() + os.sep + os.path.join('test', 'test_run_experiments', 'test_run_sqlite_experiment.py')
     assert entries[0][11] == ('Traceback (most recent call last):\n  File'
                               f' "{first_cwd_mention}"'
-                              ', line 373, in _execution_wrapper\n    experiment_function(keyfield_values, result_processor,'
+                              ', line 370, in _execution_wrapper\n    experiment_function(keyfield_values, result_processor,'
                               f' custom_fields)\n  File "{second_cwd_mention}", line 83, in error_function\n    '
                               'raise Exception("Error with weird symbos \'@#$%&/\\()=")\nException: Error with weird symbos \'@#$%&/\\()=\n')
 
