### 2023-03-31 456d707
Use joblib for multiprocessing.
Remove old dependencies, notebooks and update testcases
