### 2023-03-31 72b4094
Changed multiprocessing to joblib (#103)
* Fix and slight modification in test workflow.
* Remove unused parameters
* Add windows to settings.json
* Update poetry to include joblib
* Use joblib for multiprocessing. Remove old dependencies, notebooks and update testcases

---------

Co-authored-by: Lukas Fehring <lukasfehring@gmail.com>
Co-authored-by: Tanja Tornede <t.tornede@ai.uni-hannover.de>
