+++ pyproject.toml
@@ -33,6 +33,7 @@ numpy = ">=1.15"
 pandas = ">=1.0"
 mysql-connector-python = ">=8.0"
 jupyterlab = "^3.5.0"
+joblib = "^1.2.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = ">=7.0"
