+++ _sources/autoapi/py_experimenter/database_connector_mysql/index.rst.txt
@@ -22,7 +22,12 @@ Classes
    Bases: :py:obj:`py_experimenter.database_connector.DatabaseConnector`
 
    .. py:attribute:: _write_to_database_separator
-      :annotation: = ', '
+      :value: "', '"
+
+      
+
+   .. py:attribute:: _prepared_statement_placeholder
+      :value: '%s'
 
       
 
@@ -47,17 +52,10 @@ Classes
    .. py:method:: _create_table(cursor, columns)
 
 
-   .. py:method:: _get_tablename_for_query()
-
-
    .. py:method:: _table_has_correct_structure(cursor, typed_fields)
 
 
    .. py:method:: _pull_open_experiment(random_order) -> Tuple[int, List, List]
-
-
-   .. py:method:: escape_sql_chars(*args)
-      :staticmethod:
 
 
    .. py:method:: random_order_string()
