+++ _sources/autoapi/py_experimenter/result_processor/index.rst.txt
@@ -28,17 +28,14 @@ Attributes
 
 
 .. py:data:: result_logger
-   
 
    
 
 .. py:data:: formatter
-   
 
    
 
 .. py:data:: file_handler
-   
 
    
 
