+++ py_experimenter/experimenter.py
@@ -3,9 +3,10 @@ import logging
 import os
 import socket
 import traceback
-from joblib import Parallel, delayed
 from typing import Callable, Dict, List
+
 import pandas as pd
+from joblib import Parallel, delayed
 
 from py_experimenter import utils
 from py_experimenter.database_connector_lite import DatabaseConnectorLITE
@@ -29,21 +30,21 @@ class PyExperimenter:
         """
         Initializes the PyExperimenter with the given information.
 
-        :param experiment_configuration_file_path: The path to the experiment configuration file. Defaults to 
+        :param experiment_configuration_file_path: The path to the experiment configuration file. Defaults to
             'config/configuration.cfg'.
         :type experiment_configuration_file_path: str, optional
-        :param database_credential_file_path: The path to the database configuration file storing the credentials 
+        :param database_credential_file_path: The path to the database configuration file storing the credentials
             for the database connection, i.e., host, user and password. Defaults to 'config/database_credentials.cfg'.
         :type database_credential_file_path: str, optional
-        :param table_name: The name of the database table, if given it will overwrite the table_name given in the 
-            `experiment_configuration_file_path`. If None, the table table name is taken from the experiment 
+        :param table_name: The name of the database table, if given it will overwrite the table_name given in the
+            `experiment_configuration_file_path`. If None, the table table name is taken from the experiment
             configuration file. Defaults to None.
         :type table_name: str, optional
-        :param database_name: The name of the database, if given it will overwrite the database_name given in the 
-            `experiment_configuration_file_path`. If None, the database name is taken from the experiment configuration 
+        :param database_name: The name of the database, if given it will overwrite the database_name given in the
+            `experiment_configuration_file_path`. If None, the database name is taken from the experiment configuration
             file. Defaults to None.
         :type database_name: str, optional
-        :param name: The name of the PyExperimenter, which will be logged in the according column in the database table. 
+        :param name: The name of the PyExperimenter, which will be logged in the according column in the database table.
             Defaults to 'PyExperimenter'.
         :type name: str, optional
         :raises InvalidConfigError: If either the experiment or database configuration are missing mandatory information.
@@ -93,18 +94,18 @@ class PyExperimenter:
 
     def get_config_value(self, section_name: str, key: str) -> str:
         """
-        Returns the value of the property of the experiment configuration identified by the given key. If the `key` 
+        Returns the value of the property of the experiment configuration identified by the given key. If the `key`
         is not contained within the section, an exception is raised.
 
         :param section_name: The name of the section containing the property whose value should be returned.
         :type section_name: str
-        :param key: The name of the key identifying the property within the given section of the experiment 
+        :param key: The name of the key identifying the property within the given section of the experiment
             configuration of which a value should be returned.
         :type key: str
-        :return: The value of the property identified by the given key within the section `section_name` of the 
+        :return: The value of the property identified by the given key within the section `section_name` of the
             experiment configuration.
         :rtype: str
-        :raises NoOptionError: If the section called `section_name` is not part of the experiment configuration, or 
+        :raises NoOptionError: If the section called `section_name` is not part of the experiment configuration, or
             the `key` is not contained within that section.
         """
         return self.config.get(section_name, key)
@@ -122,15 +123,15 @@ class PyExperimenter:
 
     def has_option(self, section_name: str, key: str) -> bool:
         """
-        Checks whether the experiment configuration contains a property identified by the given 'key' within the 
+        Checks whether the experiment configuration contains a property identified by the given 'key' within the
         section called 'section_name'.
 
[…]
