+++ docs/source/usage.md
@@ -103,7 +103,7 @@ from py_experimenter.experimenter import PyExperimenter
 
 experimenter = PyExperimenter()
 experimenter.fill_table_from_config()
-experimenter.execute(run_experiment, max_experiments=-1, random_order=True)
+experimenter.execute(run_experiment, max_experiments=-1)
 ```
 
 ### Creating a PyExperimenter
@@ -158,14 +158,12 @@ An experiment can be executed easily with the following call:
 ```python
 experimenter.execute(
     experiment_function = run_experiment, 
-    max_experiments = -1, 
-    random_order = True
+    max_experiments = -1
 )
 ```
 
 - `experiment_function` is the [experiment funtion](#defining-the-experiment-function) described above.
 - `max_experiments` determines how many experiments will be executed by this `PyExperimenter`. If set to `-1`, it will execute experiments in a sequential fashion until no more open experiments are available.
-- `random_order` determines if the order in which experiments are selected for execution should be random. This is especially important to be turned on, if the execution is parallelized, e.g. on an HPC cluster.  
 
 ### Reset Experiments
 
