+++ _sources/autoapi/py_experimenter/experimenter/index.rst.txt
@@ -37,18 +37,18 @@ Classes
 
    .. py:method:: get_config_value(section_name: str, key: str) -> str
 
-      Returns the value of the property of the experiment configuration identified by the given key. If the `key` 
+      Returns the value of the property of the experiment configuration identified by the given key. If the `key`
       is not contained within the section, an exception is raised.
 
       :param section_name: The name of the section containing the property whose value should be returned.
       :type section_name: str
-      :param key: The name of the key identifying the property within the given section of the experiment 
+      :param key: The name of the key identifying the property within the given section of the experiment
           configuration of which a value should be returned.
       :type key: str
-      :return: The value of the property identified by the given key within the section `section_name` of the 
+      :return: The value of the property identified by the given key within the section `section_name` of the
           experiment configuration.
       :rtype: str
-      :raises NoOptionError: If the section called `section_name` is not part of the experiment configuration, or 
+      :raises NoOptionError: If the section called `section_name` is not part of the experiment configuration, or
           the `key` is not contained within that section.
 
 
@@ -64,15 +64,15 @@ Classes
 
    .. py:method:: has_option(section_name: str, key: str) -> bool
 
-      Checks whether the experiment configuration contains a property identified by the given 'key' within the 
+      Checks whether the experiment configuration contains a property identified by the given 'key' within the
       section called 'section_name'.
 
-      :param section_name: The name of the section of the experiment configuration of which the `key` should be 
+      :param section_name: The name of the section of the experiment configuration of which the `key` should be
           checked.
       :type section_name: str
       :param key: The name of the key to check within the given section.
       :type key: str
-      :return: True if the given `key` is contained in the experiment configuration within the section called 
+      :return: True if the given `key` is contained in the experiment configuration within the section called
           `section_name`. False otherwise.
       :rtype: bool
 
@@ -85,7 +85,7 @@ Classes
 
       :param _config: The experiment configuration.
       :type _config: configparser
-      :param database_credential_file_path: The path to the database configuration file, i.e., the file defining 
+      :param database_credential_file_path: The path to the database configuration file, i.e., the file defining
           the host, user and password. Defaults to None.
       :type database_credential_file_path: str, optional
       :return: True if the experiment configuration contains all necessary fields.
@@ -96,15 +96,15 @@ Classes
 
       Adds rows to the database table based on the given information.
 
-      First the existence of the database table is checked. If it does not exist, the database table is created 
-      based on the information in the experiment configuration file the `PyExperimenter` has been initialized 
-      with. 
+      First the existence of the database table is checked. If it does not exist, the database table is created
+      based on the information in the experiment configuration file the `PyExperimenter` has been initialized
+      with.
 
-      Afterwards, the database table is filled. To this end, the cartesian product of all `parameters` and the 
+      Afterwards, the database table is filled. To this end, the cartesian product of all `parameters` and the
       `fixed_parameter_combinations` is built, where each combination will make up a row in the database table.
       Note that only rows are added whose parameter combinations do not already exist in the database table.
-      For each added row the status is set to 'created'. If any parameter of the combinations (rows) does not 
-      match the keyfields from the experiment configuration, an error is raised. 
+      For each added row the status is set to 'created'. If any parameter of the combinations (rows) does not
+      match the keyfields from the experiment configuration, an error is raised.
 
       In the following, an example call of this method is given:
 
@@ -131,7 +131,7 @@ Classes
       :type fixed_parameter_combinations: List[dict], optional
[…]
