+++ test/test_run_experiments/test_run_mysql_experiment.py
@@ -107,12 +107,8 @@ def test_run_error_experiment():
     assert entries[0][6] == 'PyExperimenter'
     assert entries[0][8] == None
     assert entries[0][9] == None
-    assert entries[0][11] == ('Traceback (most recent call last):\n  File'
-                              f' "{os.path.join(os.getcwd(),"py_experimenter","experimenter")}.py"'
-                              ', line 376, in _execution_wrapper\n    experiment_function(keyfield_values, result_processor,'
-                              f' custom_fields)\n  File "{os.path.join(os.getcwd(), "test", "test_run_experiments", "test_run_mysql_experiment")}'
-                              '.py", line 81, in error_function\n    '
-                              'raise Exception("Error with weird symbos \'@#$%&/\\()=")\nException: Error with weird symbos \'@#$%&/\\()=\n')
+    for message in ["in _execution_wrapper", "experiment_function(keyfield_values, result_processor, custom_fields)", "raise Exception(", "Error with weird symbos \'@#$%&/\\()="]:
+        assert message in entries[0][11] 
 
 
 def own_function_raising_errors(keyfields: dict, result_processor: ResultProcessor, custom_fields: dict):
