### 2023-03-31 0dffa3e
Improve documentation of logtables and logtables example
