+++ docs/source/usage.md
@@ -1,235 +1,0 @@
-# Usage
-
-## Experiment Configuration File
-
-The experiment configuration file is primarily used to define the database backend, as well as execution parameters, i.e. keyfields, and result fields.
-
-```
-[PY_EXPERIMENTER]
-provider = sqlite 
-database = database_name
-table = table_name 
-
-keyfields = dataset, cross_validation_splits:int, seed:int, kernel
-dataset = iris
-cross_validation_splits = 5
-seed = 2:10:2 
-kernel = linear, poly, rbf, sigmoid
-
-n_jobs = 5 
-
-resultfields = pipeline:LONGTEXT, train_f1:DECIMAL, train_accuracy:DECIMAL, test_f1:DECIMAL, test_accuracy:DECIMAL
-resultfields.timestamps = false
-
-[CUSTOM] 
-path = sample_data
-```
-
-- `provider`: Either `sqlite` or `mysql`. In case of `mysql` an additional [database credential file](#database-credential-file) has to be created.
-- `database`: The name of the database.
-- `table`: The name of the table to write the experiment information into.
-- `keyfields`: The columns of the table that will define the execution of the experiments. Optionally, the field types can be attached. For each keyfield, an additional entry to the config file with the same name has to be added, which defines the domain of the keyfield.
-- `resultfields`: The columns of the table that will store the results of the execution of the experiments. Optionally, the field types can be attached.
-- `resultfields.timestamps`: Boolean defining if additional timestamp columns are needed for each resultfield.
-
-Both keyfields and resultfields can have further annotations for the data type. This is done by attaching `:<TYPE>` to the according fields. If no data type is explicitly specified, `VARCHAR(255)` is used.
-
-Keyfields of the experiment configuration, that do not have to be explicitly defined in the list of keyfields, are:
-
-- `n_jobs (int)`: The maximum number of experiments that will be executed in parallel.
-
-Additionally, the user can define which values the keyfields can take on. Usually this is done with a comma separated list of strings or numbers. In the example above, the key field `kernel` can be any of the four given values: `linear`, `poly`, `rbf`, or `sigmoid`. Note that strings are neither allowed to contain any quotation marks nor whitespace.
-
-As this manual definition can be a tedious task, especially for a list of integers, there is the option to define the start and the end of the list, together with the step size in the form: `start:end:stepsize`. In the example above, `seed` is meant to be `2, 4, 6, 8, 10`, but instead of the explicit list `2:10:2` is given.
-
-Optionally, custom configurations can be defined under the `CUSTOM` section, which will be ignored when creating or filling the database, but can provide fixed parameters for the actual execution of experiments. A common example is the path to some folder in which the data is located.
-
----
-
-## Database Credential File
-
-When working with `MySQL` as a database provider, an additional database credential file is needed, containing the credentials for accessing the database:
-
-```
-[CREDENTIALS]
-host = <host>
-user = <user>
-password = <password>
-```
-
-By default, this file is located at `config/database_credentials.cfg`. If this is not the case, the according path has to be explicitly given when [executing `PyExperimenter`](#executing-the-pyexperimenter).
-
----
-
-## Defining the Experiment Function
-
-The execution of a single experiment has to be defined within a function. The function is called with the `keyfields` of a database entry. The results are meant to be processed to be written into the database, i.e. as `resultfields`.
-
-```python
-import os
-from py_experimenter.result_processor import ResultProcessor
-
-def run_experiment(keyfields: dict, result_processor: ResultProcessor, custom_fields: dict):
-    # Extracting given parameters
-    seed = keyfields['seed']
-    path = os.path.join(custom_fields['path'], keyfields['dataset'])
-
-    # Do some stuff here
-
[…]
