+++ AUTHORS.rst
@@ -5,10 +5,10 @@ Authors
 Core Development Team
 ---------------------
 
-- `Tanja Tornede <mailto:tanja.tornede@upb.de?subject=[PyExperimenter]/>`_
-- `Alexander Tornede <mailto:alexander.tornede@upb.de?subject=[PyExperimenter]/>`_
+- `Tanja Tornede <mailto:t.tornede@ai.uni-hannover.de?subject=[PyExperimenter]/>`_
+- `Alexander Tornede <mailto:a.tornede@ai.uni-hannover.de?subject=[PyExperimenter]/>`_
 - `Lukas Fehring <mailto:fehring2@mail.uni-paderborn.de?subject=[PyExperimenter]/>`_
-- `Helena Graf <mailto:helena.graf@upb.de?subject=[PyExperimenter]/>`_
+- `Helena Graf <mailto:h.graf@ai.uni-hannover.de?subject=[PyExperimenter]/>`_
 - `Jonas Hanselle <mailto:jonas.hanselle@upb.de?subject=[PyExperimenter]/>`_
 - `Felix Mohr <mailto:felix.mohr@unisabana.edu.co?subject=[PyExperimenter]/>`_
 - `Marcel Wever <mailto:marcel.wever@ifi.lmu.de?subject=[PyExperimenter]/>`_

[…]
