### 2023-04-02 4ffb5c5
Added logtable functionality (#84)
* Added Logtable functionality, example notebook and documentation
* Reworked how tables are created in general
* Reworked large part of usage documentation
* Added timestamps to result tables
* Unified timestamp handling
* Reorder functionality in utils.py
* Added freezegun to poetry develop environment for easier testing
* Added missing path checks of examples

---------

Co-authored-by: Tanja Tornede <t.tornede@ai.uni-hannover.de>
Co-authored-by: Alexander Tornede <a.tornede@ai.uni-hannover.de>
