+++ py_experimenter/utils.py
@@ -1,11 +1,11 @@
-import configparser
 import logging
+from configparser import ConfigParser
 from datetime import datetime
-from typing import List, Tuple
+from typing import Dict, List, Optional, Tuple
 
 import numpy as np
 
-from py_experimenter.exceptions import ConfigError, NoConfigFileError, ParameterCombinationError
+from py_experimenter.exceptions import ConfigError, MissingLogTableError, NoConfigFileError, ParameterCombinationError
 
 
 def load_config(path):
@@ -14,7 +14,7 @@ def load_config(path):
     :param path: path to the config file
     :return: configuration file
     """
-    config = configparser.ConfigParser()
+    config = ConfigParser()
     try:
         with open(path) as f:
             config.read_file(f)
@@ -22,44 +22,6 @@ def load_config(path):
         raise NoConfigFileError(f'Configuration file missing! Please add file: {path}')
 
     return config
-
-
-def timestamps_for_result_fields(config: configparser.ConfigParser) -> bool:
-    if config.has_option('PY_EXPERIMENTER', 'resultfields.timestamps'):
-        timestamp_on_result_fields = config.getboolean('PY_EXPERIMENTER', 'resultfields.timestamps')
-    else:
-        timestamp_on_result_fields = False
-    return timestamp_on_result_fields
-
-
-def add_timestep_result_columns(result_field_configuration):
-    result_fields_with_timestamp = list()
-    for result_field in result_field_configuration:
-        result_fields_with_timestamp.append(result_field)
-        result_fields_with_timestamp.append((f'{result_field[0]}_timestamp', 'VARCHAR(255)'))
-    return result_fields_with_timestamp
-
-
-def extract_db_credentials_and_table_name_from_config(config):
-    """
-    Initialize connection to database based on configuration file. If the tables does not exist, a new one will be
-    created automatically
-    :param config: Configuration file with database and experiment information
-    :return: mysql_connector and table name from the config file
-    """
-    database_config = config['PY_EXPERIMENTER']
-    if database_config['provider'] == 'sqlite':
-        host = None
-        user = None
-        password = None
-    else:
-        host = database_config['host']
-        user = database_config['user']
-        password = database_config['password']
-    database = database_config['PY_EXPERIMENTER']
-    table_name = database_config['table'].replace(' ', '')
-
-    return table_name, host, user, database, password
 
 
 def get_keyfield_data(config):
@@ -85,66 +47,49 @@ def get_keyfield_data(config):
     return keyfield_data
 
 
-def _generate_int_data(keyfield_values):
-    final_data = []
-    for data_definition in keyfield_values:
-        if ':' in data_definition:
+def extract_db_credentials_and_table_name_from_config(config):
+    """
+    Initialize connection to database based on configuration file. If the tables does not exist, a new one will be
+    created automatically
+    :param config: Configuration file with database and experiment information
+    :return: mysql_connector and table name from the config file
+    """
+    database_config = config['PY_EXPERIMENTER']
+    if database_config['provider'] == 'sqlite':
+        host = None
+        user = None
+        password = None
+    else:
+        host = database_config['host']
+        user = database_config['user']
+        password = database_config['password']
+    database = database_config['PY_EXPERIMENTER']
+    table_name = database_config['table'].replace(' ', '')
 
-            if data_definition.startswith(':') or data_definition.endswith(':') or '::' in data_definition:
-                raise ConfigError(f'{data_definition} is not a valid integer range')
-
-            integer_range = data_definition.split(':')
-
[…]
