+++ _sources/examples/example_conditional_grid.ipynb.txt
@@ -1,12 +1,13 @@
 {
 	"cells": [
 		{
+			"attachments": {},
 			"cell_type": "markdown",
 			"metadata": {
 				"id": "iLnG0-_UC6ig"
 			},
 			"source": [
-				"# Conditional Parameter Grids\n",
+				"# Example: Conditional Parameter Grids\n",
 				"\n",
 				"This example shows the usage of `PyExperimenter` with a conditional parameter grid. We will programmatically define the parameter combinations of a support vector machine, instead of generating the entire cartesian product from the parameters defined in the config file.  \n",
 				"\n",
@@ -51,6 +52,12 @@
 				"[CUSTOM] \n",
 				"path = sample_data\n",
 				"\"\"\"\n",
+				"\n",
+				"# Create config directory if it does not exist\n",
+				"if not os.path.exists('config'):\n",
+				"    os.mkdir('config')\n",
+				"    \n",
+				"# Create config file\n",
 				"experiment_configuration_file_path = os.path.join('config', 'example_conditional_grid.cfg')\n",
 				"with open(experiment_configuration_file_path, \"w\") as f: \n",
 				"  f.write(content)"
@@ -255,7 +262,7 @@
 							"      <td>0.1</td>\n",
 							"      <td>nan</td>\n",
 							"      <td>nan</td>\n",
-							"      <td>2023-03-31 17:29:52</td>\n",
+							"      <td>2023-03-31 20:48:29</td>\n",
 							"      <td>created</td>\n",
 							"      <td>None</td>\n",
 							"      <td>None</td>\n",
@@ -277,7 +284,7 @@
 							"      <td>0.3</td>\n",
 							"      <td>nan</td>\n",
 							"      <td>nan</td>\n",
-							"      <td>2023-03-31 17:29:52</td>\n",
+							"      <td>2023-03-31 20:48:29</td>\n",
 							"      <td>created</td>\n",
 							"      <td>None</td>\n",
 							"      <td>None</td>\n",
@@ -299,7 +306,7 @@
 							"      <td>0.1</td>\n",
 							"      <td>3</td>\n",
 							"      <td>0</td>\n",
-							"      <td>2023-03-31 17:29:52</td>\n",
+							"      <td>2023-03-31 20:48:29</td>\n",
 							"      <td>created</td>\n",
 							"      <td>None</td>\n",
 							"      <td>None</td>\n",
@@ -321,7 +328,7 @@
 							"      <td>0.1</td>\n",
 							"      <td>3</td>\n",
 							"      <td>0.1</td>\n",
-							"      <td>2023-03-31 17:29:52</td>\n",
+							"      <td>2023-03-31 20:48:29</td>\n",
 							"      <td>created</td>\n",
 							"      <td>None</td>\n",
 							"      <td>None</td>\n",
@@ -343,7 +350,7 @@
 							"      <td>0.1</td>\n",
 							"      <td>4</td>\n",
 							"      <td>0</td>\n",
-							"      <td>2023-03-31 17:29:52</td>\n",
+							"      <td>2023-03-31 20:48:29</td>\n",
 							"      <td>created</td>\n",
 							"      <td>None</td>\n",
 							"      <td>None</td>\n",
@@ -365,7 +372,7 @@
 							"      <td>0.1</td>\n",
 							"      <td>4</td>\n",
 							"      <td>0.1</td>\n",
-							"      <td>2023-03-31 17:29:52</td>\n",
+							"      <td>2023-03-31 20:48:29</td>\n",
[…]
