+++ _sources/autoapi/py_experimenter/exceptions/index.rst.txt
@@ -56,7 +56,7 @@ Module Contents
    Common base class for all non-exit exceptions.
 
 
-.. py:exception:: TableError
+.. py:exception:: CreatingTableError
 
    Bases: :py:obj:`DatabaseError`
 
@@ -72,7 +72,7 @@ Module Contents
 
 .. py:exception:: TableHasWrongStructureError
 
-   Bases: :py:obj:`TableError`
+   Bases: :py:obj:`CreatingTableError`
 
    Common base class for all non-exit exceptions.
 
@@ -106,6 +106,13 @@ Module Contents
 
 
 .. py:exception:: InvalidValuesInConfiguration
+
+   Bases: :py:obj:`ConfigError`
+
+   Common base class for all non-exit exceptions.
+
+
+.. py:exception:: MissingLogTableError
 
    Bases: :py:obj:`ConfigError`
 
