+++ _sources/autoapi/py_experimenter/result_processor/index.rst.txt
@@ -51,11 +51,14 @@ Attributes
       :param results: Dictionary with result field name and result value pairs.
 
 
-   .. py:method:: _add_timestamps_to_results(results: dict, time: datetime.datetime) -> List[Tuple[str, object]]
+   .. py:method:: _add_timestamps_to_results(results: dict) -> List[Tuple[str, object]]
       :staticmethod:
 
 
-   .. py:method:: _change_status(status)
+   .. py:method:: process_logs(logs: Dict[str, Dict[str, str]]) -> None
+
+
+   .. py:method:: _change_status(status: str)
 
 
    .. py:method:: _write_error(error_msg)

[…]
