### 2023-04-02 39ebb0e
Updated documentation for resetting experiments
