### 2023-04-02 fd99724
Removed deprecated documentation sentence
