### 2023-04-02 0eee2e5
Fixed documentation of reset experiments (#106)
