### 2023-04-02 bcdaf11
Removed duplicated wording
