### 2023-04-02 tornede/py_experimenter#108 @tornede
Dockerfile
[One reviewer of the JOSS Paper provided a minimalistic Dockerfile](https://github.com/openjournals/joss-reviews/issues/5149#issuecomment-1467999919) which could be added to the repository.

```
cd
mkdir py_experimenter
cd py_experimenter
mkdir Docker
vim Docker/Dockerfile
```

Content added to Dockerfile:

```
FROM ubuntu:latest
RUN apt-get update && apt-get upgrade -y
RUN apt-get install python3-pip -y
RUN pip3 install py-experimenter
RUN pip3 install scikit-learn

WORKDIR /py_experimenter
```

Build and enter the container:

```
docker build Docker -t py_experimenter_docker
docker run -it -v ~/py_experimenter_test:/py_experimenter py_experimenter_docker
```
