### 2023-04-02 9347c15
Removed uplicated line of code
