+++ AUTHORS.rst
@@ -7,7 +7,7 @@ Core Development Team
 
 - `Tanja Tornede <mailto:t.tornede@ai.uni-hannover.de?subject=[PyExperimenter]/>`_
 - `Alexander Tornede <mailto:a.tornede@ai.uni-hannover.de?subject=[PyExperimenter]/>`_
-- `Lukas Fehring <mailto:fehring2@mail.uni-paderborn.de?subject=[PyExperimenter]/>`_
+- `Lukas Fehring <mailto:lukas.fehring@stud.uni-hannover.de?subject=[PyExperimenter]/>`_
 - `Helena Graf <mailto:h.graf@ai.uni-hannover.de?subject=[PyExperimenter]/>`_
 - `Jonas Hanselle <mailto:jonas.hanselle@upb.de?subject=[PyExperimenter]/>`_
 - `Felix Mohr <mailto:felix.mohr@unisabana.edu.co?subject=[PyExperimenter]/>`_
