+++ CHANGELOG.rst
@@ -9,13 +9,13 @@ v1.2.0 (03/04/2023)
 Feature
 -------
 
-- Added logtables functionality, allowing to log information during the execution of an experiment, which is described in detail in the documentation.
+- Added logtables functionality, allowing to incrementally log information during the execution of an experiment, which is described in detail in the documentation.
 - Documentation of the usage of ``PyExperimenter`` has been reworked in large parts. 
 
 Examples
 --------
 - An additional logtable example has been added.
-- An issue of the example notebook has been fixed causing them to fail due to missing folder. 
+- An issue of the example notebook has been fixed causing them to fail due to missing directories. 
 - Improved general example to cover extended functionality of ``PyExperimenter.reset_experiments()``.
 
 Fix
@@ -23,8 +23,8 @@ Fix
 
 - Start date is now set when pulling an experiment
 - Supported Python version is now >= 3.9
-- Changed row identification in ResultProcessor to ExperimentID instead of checking keyfields
-- Stack traces is now correctly logged into mysql database, thereby the used mysql connector implementation has been changed to C 
+- Changed row identification in ResultProcessor to experiment ID instead of checking keyfields
+- Stack traces are now correctly logged into the mysql database, as the used mysql connector implementation has been changed to C. 
 - Changed multiprocessing to joblib due to issues with the example notebooks
 - The ``random_order`` parameter is not needed anymore for the execution, therefore it has been removed. 
 - Documentation of ``PyExperimenter.reset_experiments()`` has been updated to reflect the changes in the functionality.
