### 2023-04-03 259edbf
Update version to 1.2.0 (#109)
* Update version and changelog
* Updated author email address
* Removed duplicated line of code
* Remove unnecessary "the" in document title

---------

Co-authored-by: Lukas Fehring <lukas.ferhing@stud.uni-hannover.de>
