+++ paper/paper.md
@@ -37,7 +37,7 @@ affiliations:
    index: 2
  - name: MCML, Institut for Informatics, LMU Munich, Germany
    index: 3
-date: 2 April 2023
+date: 4 April 2023
 bibliography: paper.bib
 
 ---
