+++ pyproject.toml
@@ -1,11 +1,11 @@
 [tool.poetry]
 name = "py-experimenter"
-version = "1.1.0"
+version = "1.2.0"
 description = "The PyExperimenter is a tool for the automatic execution of experiments, e.g. for machine learning (ML), capturing corresponding results in a unified manner in a database."
 authors = [
     "Tanja Tornede <t.tornede@ai.uni-hannover.de>",
     "Alexander Tornede <a.tornede@ai.uni-hannover.de>",
-    "Lukas Fehring <fehring2@mail.uni-paderborn.de>",
+    "Lukas Fehring <lukas.fehring@stud.uni-hannover.de>",
     "Lukas Gehring",
     "Helena Graf <h.graf@ai.uni-hannover.de>",
     "Jonas Hanselle <jonas.hanselle@upb.de>",
