### 2023-04-04 28cb35f
Improve sentence style within documentation
