+++ _sources/autoapi/py_experimenter/utils/index.rst.txt
@@ -32,7 +32,7 @@ Functions
 
 .. py:function:: load_config(path)
 
-   Load and return configuration file
+   Load and return configuration file.
    :param path: path to the config file
    :return: configuration file
 
@@ -43,7 +43,7 @@ Functions
 .. py:function:: extract_db_credentials_and_table_name_from_config(config)
 
    Initialize connection to database based on configuration file. If the tables does not exist, a new one will be
-   created automatically
+   created automatically.
    :param config: Configuration file with database and experiment information
    :return: mysql_connector and table name from the config file
 
