### 2023-04-04 #5149 @tornede
Dear reviewers and editor,

thanks a lot (again) for the kind and detailed feedback! It helped a lot to find outdated documentation and issues within the examples.

`PyExperimenter` has been developed further since we handed in the paper in November 2022. The last week, we concluded a lot of unfinished work and prepared a new release, both addressing your issues as well as new features. Additionally we reworked huge parts of the documentation, especially regarding usage, and updated the example notebooks.

# @schnorr
> I've only found a minor problem that I've registered as [an issue in their repository](https://github.com/tornede/py_experimenter/issues/90#issue-1620357634).

Thanks for the detailed issue regarding resetting experiments, which we fixed within the last days.

Additionally you checked all boxes from the review checklist, which is why we assume that you have no further concerns. :)

# @ArsamAryandoust

> A list of dependencies is available in the pyproject.toml file, maybe could also be stated in the readme or a requirements.txt file (?).

We decided to use poetry, and therefore define dependencies via [pyproject.toml](https://github.com/tornede/py_experimenter/blob/main/pyproject.toml). Having a separated requirements file would be duplicated and error prone. We decided to reduce the README to a minimum and instead reference the documentation, again to reduce redundancy and possible errors.

> 1. The documentations under docs/source/examples/ need to be improved. Test your notebooks more accurately and make sure they run. I made exemplar pull requests. Everything else works flawlessly.

Thanks for hinting to the problem of the creation of the experiment configuration file within the example notebooks, which we fixed within the last days.

> 2. Seeking support and creating issues is described under documents tab 'Need help?'. Descriptions on how to contribute can be added.

We distinguish between a simple user and an (external) contributor, which is why we have the two pages within the documentation: "Need Help?" and "How to Contribute". In case of a simple user and any occurring problems, we believe referring to the examples and GitHub issues is enough, as such a person will most likely not make actual changes to the code. Otherwise "How to Contribute" can be found next to "Need Help?" within the side bar.

> 3. I did not test the full functionality of the presented tool. However, the provided toy examples are functional and seem to demonstrate the core functionality.

Thanks for the great feedback, that the provided examples demonstrate the core functionality. As this has been listed under "issues to be addressed" we are wondering, if you have any further feedback that can be addressed?

> The paper can have small improvements.

Thanks for the feedback to the paper text itself. We implemented some parts of it into the paper.

> I like working with Docker. So I create a minimalistic Dockerfile

Thanks a lot providing a Dockerfile and some lines how to build it. We created an [issue](https://github.com/tornede/py_experimenter/issues/108) containing your provided information, which will be investigated in the future.

> In example notebooks, there are small issues.

Thanks for pointing this out, we fixed it within the last days.

> In code snippet under 'Fill Table' section, change code into the following, to ensure that code runs while also illustrating the example. I guess everyone will still get the point also with dataset being iris again

Thanks for the remark, that this row might not be well understood. We decided to change the dataset name to "error_dataset", which will always result into a raised exception and therefore a logged stacktrace.

> In code section under 'Restart Failed Experiments', change code into [...]

Thanks for this issue, which is the same as pointed out by @schnorr, which has been fixed within the last days.

When checking your review checklist, we figured out that some boxes have not been checked. Is there anything else that has to be addressed?

# @timtroendle
How do we handle the new version for the JOSS paper? Will it be the old version v1.1.0 or the new one v1.2.0 ?
