### 2023-04-06 ed529af
initial changes for sqlite and general connector
