+++ py_experimenter/database_connector.py
@@ -1,23 +1,25 @@
 import abc
 import logging
-from datetime import datetime
-from typing import List, Optional, Tuple
+from configparser import ConfigParser
+from typing import Dict, List, Optional, Tuple, Union
 
 import pandas as pd
 
 from py_experimenter import utils
-from py_experimenter.exceptions import DatabaseConnectionError, EmptyFillDatabaseCallError, NoExperimentsLeftException, TableHasWrongStructureError
+from py_experimenter.exceptions import (CreatingTableError, DatabaseConnectionError, EmptyFillDatabaseCallError, NoExperimentsLeftException,
+                                        TableHasWrongStructureError)
 from py_experimenter.experiment_status import ExperimentStatus
+
 
 class DatabaseConnector(abc.ABC):
 
-    def __init__(self, database_credential_file_path):
-        self.database_credential_file_path = database_credential_file_path
-        self.table_name = self.database_credential_file_path.get('PY_EXPERIMENTER', 'table')
-        self.database_name = self.database_credential_file_path.get('PY_EXPERIMENTER', 'database')
+    def __init__(self, config: ConfigParser):
+        self.config = config
+        self.table_name = self.config.get('PY_EXPERIMENTER', 'table')
+        self.database_name = self.config.get('PY_EXPERIMENTER', 'database')
 
         self.database_credentials = self._extract_credentials()
-        self.timestamp_on_result_fields = utils.timestamps_for_result_fields(self.database_credential_file_path)
+        self.timestamp_on_result_fields = utils.timestamps_for_result_fields(self.config)
 
         self._test_connection()
 
@@ -66,63 +68,75 @@ class DatabaseConnector(abc.ABC):
     def create_table_if_not_existing(self) -> None:
         logging.debug("Create table if not exist")
 
-        keyfields = utils.get_keyfields(self.database_credential_file_path)
-        resultfields = utils.get_resultfields(self.database_credential_file_path)
-
+        keyfields = utils.get_keyfields(self.config)
+        resultfields = utils.get_resultfields(self.config)
         if self.timestamp_on_result_fields:
             resultfields = utils.add_timestep_result_columns(resultfields)
 
         connection = self.connect()
         cursor = self.cursor(connection)
-
         if self._table_exists(cursor):
             if not self._table_has_correct_structure(cursor, keyfields + resultfields):
                 raise TableHasWrongStructureError("Keyfields or resultfields from the configuration do not match columns in the existing "
                                                   "table. Please change your configuration or delete the table in your database.")
         else:
-            fields = (
-                keyfields +
-                [
-                    ('creation_date', 'VARCHAR(255)'),
-                    ('status', 'VARCHAR(255)'),
-                    ('start_date', 'VARCHAR(255)'),
-                    ('name', 'LONGTEXT'),
-                    ('machine', 'VARCHAR(255)')
-                ]
-                + resultfields +
-                [
-                    ('end_date', 'VARCHAR(255)'),
-                    ('error', 'LONGTEXT'),
-                ]
-            )
+            columns = self._compute_columns(keyfields, resultfields)
+            self._create_table(cursor, columns, self.table_name)
 
-            columns = ['%s %s DEFAULT NULL' % (self.__class__.escape_sql_chars(field)[0], datatype) for field, datatype in fields]
-            self._create_table(cursor, columns)
+            for logtable_name, logtable_columns in utils.extract_logtables(self.config, self.table_name).items():
+                self._create_table(cursor, logtable_columns, logtable_name, logtable=True)
+
[…]
