### 2023-04-06 059b43c
adjusted test to the new interface definition of _write_database
