### 2023-04-11 8f52e5d
two fixes:
inserted transactions again
added uncommented code again
