+++ test/test_logtables/test_mariadb.py
@@ -1,0 +1,113 @@
+import os
+from configparser import ConfigParser
+from math import cos, sin
+
+from freezegun import freeze_time
+from mock import call, patch
+
+from py_experimenter.database_connector import DatabaseConnector
+from py_experimenter.database_connector_mariadb import DatabaseConnectorMariaDB
+from py_experimenter.experimenter import PyExperimenter
+from py_experimenter.result_processor import ResultProcessor
+
+
+@patch('py_experimenter.experimenter.DatabaseConnectorMYSQL._create_database_if_not_existing')
+@patch('py_experimenter.experimenter.DatabaseConnectorMYSQL._test_connection')
+@patch('py_experimenter.experimenter.DatabaseConnectorMYSQL.fill_table')
+@patch('py_experimenter.experimenter.DatabaseConnectorMYSQL.connect')
+@patch('py_experimenter.experimenter.DatabaseConnectorMYSQL.cursor')
+@patch('py_experimenter.experimenter.DatabaseConnectorMYSQL.fetchall')
+@patch('py_experimenter.experimenter.DatabaseConnectorMYSQL.close_connection')
+@patch('py_experimenter.experimenter.DatabaseConnectorMYSQL.execute')
+def test_tables_created(execute_mock, close_connection_mock, fetchall_mock, cursor_mock, connect_mock, fill_table_mock, test_connection_mock, create_database_mock):
+    execute_mock.return_value = None
+    fetchall_mock.return_value = None
+    close_connection_mock.return_value = None
+    cursor_mock.return_value = None
+    connect_mock.return_value = None
+    fill_table_mock.return_value = None
+    create_database_mock.return_value = None
+    test_connection_mock.return_value = None
+    experimenter = PyExperimenter(os.path.join('test', 'test_logtables', 'mariadb_logtables.cfg'))
+    experimenter.fill_table_from_config()
+    assert execute_mock.mock_calls[1][1][1] == ('CREATE TABLE test_mariadb_logtables (ID INTEGER PRIMARY KEY AUTO_INCREMENT, value int DEFAULT NULL,'
+                                                'exponent int DEFAULT NULL,creation_date DATETIME DEFAULT NULL,status VARCHAR(255) DEFAULT NULL,'
+                                                'start_date DATETIME DEFAULT NULL,name LONGTEXT DEFAULT NULL,machine VARCHAR(255) DEFAULT NULL,'
+                                                'sin VARCHAR(255) DEFAULT NULL,cos VARCHAR(255) DEFAULT NULL,end_date DATETIME DEFAULT NULL,'
+                                                'error LONGTEXT DEFAULT NULL);')
+    print(execute_mock.mock_calls[2][1][1])
+    assert execute_mock.mock_calls[2][1][1] == ('CREATE TABLE test_mariadb_logtables__test_mariadb_log (ID INTEGER PRIMARY KEY AUTO_INCREMENT,'
+                                                ' experiment_id INTEGER, timestamp DATETIME, test int DEFAULT NULL, FOREIGN KEY (experiment_id)'
+                                                ' REFERENCES test_mariadb_logtables(ID) ON DELETE CASCADE);')
+
+
+@freeze_time("2012-01-14 03:21:34")
+@patch('py_experimenter.result_processor.DatabaseConnectorMariaDB')
+def test_logtable_insertion(database_connector_mock):
+    fixed_time = '2012-01-14 03:21:34'
+    config = ConfigParser()
+    config.read(os.path.join('test', 'test_logtables', 'mariadb_logtables.cfg'))
+    result_processor = ResultProcessor(config, None, None, None, 0)
+    result_processor._table_name = 'some_table_name'
[…]
