### 2023-04-11 5a73b72
Update tests.yml (#114)
ignoring mariadb tests

Co-authored-by: tornede <47322192+tornede@users.noreply.github.com>
