+++ _sources/autoapi/py_experimenter/database_connector_lite/index.rst.txt
@@ -17,9 +17,12 @@ Classes
 
 
 
-.. py:class:: DatabaseConnectorLITE
+.. py:class:: DatabaseConnectorLITE(config: configparser.ConfigParser)
 
    Bases: :py:obj:`py_experimenter.database_connector.DatabaseConnector`
+
+   Helper class that provides a standard way to create an ABC using
+   inheritance.
 
    .. py:attribute:: _write_to_database_separator
       :value: "','"
